+++ data/i18n/hydrogen_es.ts
@@ -1912,6 +1912,21 @@ Please set your system&apos;s locale to UTF-8!</source>
         <extracomment>Error dialog shown when connecting to a remote headless engine fails. * The endpoint is appended via concatenation.</extracomment>
         <translation type="unfinished"></translation>
     </message>
+    <message>
+        <source>Hydrogen Project</source>
+        <extracomment>Used as file type filter for .h2project files.</extracomment>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Export Project</source>
+        <extracomment>Menu action to export the current song as a self-contained .h2project * bundle (song + embedded samples).</extracomment>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Unable to export project</source>
+        <extracomment>Shown in a dialog on .h2project export failure.</extracomment>
+        <translation type="unfinished"></translation>
+    </message>
 </context>
 <context>
     <name>ComponentEditor</name>
@@ -2866,6 +2881,10 @@ Debería funcionar correctamente mientras utilices el GMRockKit, y no uses tresi
     </message>
     <message>
         <source>&amp;Duplicate</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Project exported to</source>
         <translation type="unfinished"></translation>
     </message>
 </context>
@@ -5569,6 +5588,10 @@ La ruta al script y al nombre del script no pueden contener espacios en blanco.<
     </message>
     <message>
         <source>Count in and start/stop playback</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Export song as self-contained .h2project bundle</source>
         <translation type="unfinished"></translation>
     </message>
 </context>